Moving TaskAwaitable to separate file.
